requirements.txt
groq==1.0.0
langchain_community==0.4.1
langchain_google_genai==4.1.2
langchain_text_splitters==1.1.0
pandas==2.3.3
protobuf==6.33.2
python-dotenv==1.2.1
PyYAML==6.0.3
scikit_learn==1.8.0
streamlit==1.52.2
tqdm==4.67.1
